change defaulf config from bf16 to fp16

+++ config/base.py
@@ -18,7 +18,7 @@ def get_config():
     # number of checkpoints to keep before overwriting old ones.
     config.num_checkpoint_limit = 5
     # mixed precision training. options are "fp16", "bf16", and "no". half-precision speeds up training significantly.
-    config.mixed_precision = "bf16"
+    config.mixed_precision = "fp16"
     # allow tf32 on Ampere GPUs, which can speed up training.
     config.allow_tf32 = True
     # whether or not to use LoRA.
